% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{astral}
\alias{astral}
\title{astral}
\usage{
astral(..., java_args = NULL)
}
\arguments{
\item{...}{Arguments for ASTRAL}

\item{java_args}{Arguments for java}
}
\description{
Run astral
}
\details{
Change memory settings with \code{-Xmx#M} in \code{java_args}.
}
\examples{
library(outsider)
astral <- module_import('astral', repo = 'dombennett/om..astral')
astral('--help')
# # Elaborate exmaple (NOT RUN)
# # download example file
# download.file(url = 'https://raw.githubusercontent.com/DomBennett/om..astral/master/test_data/song_mammals.424.gene.tre')
# # run with test file and 1GB memmory
# astral('-i', 'song_mammals.424.gene.tre', '-o', 'song_mammals.tre', java_args = '-Xmx1000M')
}
