% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{astral}
\alias{astral}
\title{astral}
\usage{
astral(...)
}
\arguments{
\item{...}{Arguments}
}
\description{
Run astral
}
\examples{
library(outsider)
astral <- module_import('astral', repo = 'dombennett/om..astral')
astral('--help')
}
